requirements.txt
Django==4.1.1
pip==22.2.2
Pillow==9.2.0
asgiref==3.5.2
psycopg2-binary==2.9.5
wheel==0.37.1
sqlparse==0.4.2
include==0.2.2
djangorestframework==3.14.0